+++ kontrol-service/engine/flow/render.go
@@ -232,12 +232,26 @@ func getDestinationRule(serviceID string, services []*resolved.Service, namespac
 	// if we do that then the render work around isn't necessary
 	subsets := lo.UniqBy(
 		lo.Map(services, func(service *resolved.Service, _ int) *v1alpha3.Subset {
-			return &v1alpha3.Subset{
+
+			newSubset := &v1alpha3.Subset{
 				Name: service.Version,
 				Labels: map[string]string{
 					"version": service.Version,
 				},
 			}
+
+			// TODO Narrow down this configuration to only subsets created for telepresence intercepts or find a way to enable TLS for telepresence intercepts https://github.com/kurtosis-tech/kardinal-kontrol/issues/14
+			// This config is necessary for Kardinal/Telepresence (https://www.telepresence.io/) integration
+			if service.Version != prodVersion {
+				newTrafficPolicy := &v1alpha3.TrafficPolicy{
+					Tls: &v1alpha3.ClientTLSSettings{
+						Mode: v1alpha3.ClientTLSSettings_DISABLE,
+					},
+				}
+				newSubset.TrafficPolicy = newTrafficPolicy
+			}
+
+			return newSubset
 		}),
 		func(subset *v1alpha3.Subset) string {
 			return subset.Name
@@ -334,7 +348,16 @@ func getGateway(ingresses []*resolved.Ingress, namespace string) *istioclient.Ga
 	}
 	extHosts = lo.Uniq(extHosts)
 
-	ingressId := ingresses[0].IngressID
+	// We need to return a gateway as part of the cluster resources so we return a dummy one
+	// if there are no ingresses defined.  This can happen when the tenant does not have a base
+	// cluster topology: no initial deploy or the topologies have been deleted.  This gateway also allows
+	// us to communicate the namespace to the kardinal manager helping the resources to be cleaned up.
+	ingressId := "dummy"
+	if len(ingresses) > 0 {
+		ingressId = ingresses[0].IngressID
+	} else {
+		extHosts = []string{"dummy.kardinal.dev"}
+	}
 
 	return &istioclient.Gateway{
 		TypeMeta: metav1.TypeMeta{